fix: 계좌내역 조회 오류 해결
